practice/chapter7/code/requirements.txt
# 7장 실습: 단일 에이전트 vs 멀티에이전트 비교
# 설치: pip install -r requirements.txt

langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.50.0
python-dotenv>=1.0.0
